requirements-dev.txt
-e .
isort>=4.2.1
pep257==0.7.0
pep8>=1.6.2
flake8>=2.4.1
pep8-naming>=0.3.3
coverage>=4.0
django-coverage-plugin==1.1
mccabe>=0.3.0
pre-commit>=0.5.5
py==1.4.31
pytest==2.9.1
pytest-django==2.9.1
mock==1.3.0
pbr==1.8.1
